Extracted an interface over MarkupEngine To be able to process images in a correct fashion, markup engine can't be the base abstraction as it requires splitting the file as lines, which jpeg images have not.
